Big Five Personality Traits Data

A dataset containing responses to 20 items measuring the Big Five personality traits. This dataset was originally collected by the package author. The items are arranged in a cyclical order representing the five factors: Extraversion, Neuroticism, Openness, Agreeableness, and Conscientiousness.

Usage

BigFive

Format

A data frame with 10 rows and 20 columns:

BF1

Extraversion item 1

BF2

Neuroticism item 1

BF3

Openness item 1

BF4

Agreeableness item 1

BF5

Conscientiousness item 1

BF6

Extraversion item 2

BF7

Neuroticism item 2

BF8

Openness item 2

BF9

Agreeableness item 2

BF10

Conscientiousness item 2

BF11

Extraversion item 3

BF12

Neuroticism item 3

BF13

Openness item 3

BF14

Agreeableness item 3

BF15

Conscientiousness item 3

BF16

Extraversion item 4

BF17

Neuroticism item 4

BF18

Openness item 4

BF19

Agreeableness item 4

BF20

Conscientiousness item 4

Source

Originally collected by the package author.